+++ processor/__init__.py
@@ -5,13 +5,11 @@ from dotenv import load_dotenv
 import mediacloud.api
 import mediacloud_legacy.api
 from flask import Flask
-#from sentry_sdk.integrations.flask import FlaskIntegration
 from sentry_sdk.integrations.logging import ignore_logger
 from sentry_sdk import init
 from typing import Dict
-from sqlalchemy import create_engine
 
-VERSION = "3.2.1"
+VERSION = "3.6.0"
 SOURCE_GOOGLE_ALERTS = "google-alerts"
 SOURCE_MEDIA_CLOUD = "media-cloud"
 SOURCE_NEWSCATCHER = "newscatcher"
@@ -29,7 +27,8 @@ logger = logging.getLogger(__name__)
 logger.info("------------------------------------------------------------------------")
 logger.info("Starting up Feminicide Story Processor v{}".format(VERSION))
 # supress annoying "not enough comments" and "using custom extraction" notes# logger = logging.getLogger(__name__)
-loggers_to_skip = ['trafilatura.core', 'trafilatura.metadata', 'readability.readability']
+loggers_to_skip = ['trafilatura.core', 'trafilatura.metadata', 'readability.readability',
+                   'trafilatura.readability_lxml', 'trafilatura.htmlprocessing', 'trafilatura.xml']
 for item in loggers_to_skip:
     logging.getLogger(item).setLevel(logging.WARNING)
 
@@ -47,7 +46,7 @@ BROKER_URL = os.environ.get('BROKER_URL', None)
 if BROKER_URL is None:
     logger.warning("  ⚠️ No BROKER_URL env var specified. Using sqlite, which will perform poorly")
     BROKER_URL = "db+sqlite:///results.sqlite"
-logger.info("  Queue at {}".format(BROKER_URL))
+#logger.info("  Queue at {}".format(BROKER_URL))
 
 SENTRY_DSN = os.environ.get('SENTRY_DSN', None)  # optional
 if SENTRY_DSN:
@@ -56,15 +55,15 @@ if SENTRY_DSN:
          integrations=[CeleryIntegration()])
     ignore_logger('trafilatura.utils')
     logger.info("  SENTRY_DSN: {}".format(SENTRY_DSN))
-else:
-    logger.info("  Not logging errors to Sentry")
+#else:
+#    logger.info("  Not logging errors to Sentry")
 
 FEMINICIDE_API_URL = os.environ.get('FEMINICIDE_API_URL', None)
 if FEMINICIDE_API_URL is None:
     logger.error("  ❌ No FEMINICIDE_API_URL is specified. Bailing because we can't list projects to run!")
     sys.exit(1)
-else:
-    logger.info("  Config server at {}".format(FEMINICIDE_API_URL))
+#else:
+#    logger.info("  Config server at at {}".format(FEMINICIDE_API_URL))
 
 FEMINICIDE_API_KEY = os.environ.get('FEMINICIDE_API_KEY', None)
 if FEMINICIDE_API_KEY is None:
@@ -74,10 +73,6 @@ if FEMINICIDE_API_KEY is None:
 SQLALCHEMY_DATABASE_URI = os.environ.get('DATABASE_URL', None)
 if SQLALCHEMY_DATABASE_URI is None:
     logger.warning("  ⚠️ ️No SQLALCHEMY_DATABASE_URI is specified. Using sqlite which will perform poorly")
-    SQLALCHEMY_DATABASE_URI = 'sqlite:///data.db'
-    engine = create_engine(SQLALCHEMY_DATABASE_URI)  # use defaults (probably in test mode)
-else:
-    engine = create_engine(SQLALCHEMY_DATABASE_URI, pool_size=20, max_overflow=20)  # bumped pool size up for parallel tasks
 
 
 ENTITY_SERVER_URL = os.environ['ENTITY_SERVER_URL']
@@ -89,8 +84,20 @@ NEWSCATCHER_API_KEY = os.environ['NEWSCATCHER_API_KEY']
 if NEWSCATCHER_API_KEY is None:
     logger.warning("  ⚠️ No NEWSCATCHER_API_KEY is specified. We won't be fetching from Newscatcher.")
 
+SLACK_APP_TOKEN = os.environ.get('SLACK_APP_TOKEN', None)
+if SLACK_APP_TOKEN is None:
+    logger.warning("  ⚠️ No SLACK_APP_TOKEN env var specified. We won't be sending slack updates.")
 
-def get_mc_client() -> mediacloud.api.DirectoryApi:
+SLACK_BOT_TOKEN = os.environ.get('SLACK_BOT_TOKEN', None)
+if SLACK_BOT_TOKEN is None:
+    logger.warning("  ⚠️ No SLACK_BOT_TOKEN env var specified. We won't be sending slack updates.")
+
+SLACK_CHANNEL_ID = os.environ.get('SLACK_CHANNEL_ID', None)
+if SLACK_CHANNEL_ID is None:
+    logger.warning("  ⚠️ No CHANNEL_ID env var specified. We won't be sending slack updates.")
+
+
+def get_mc_directory_client() -> mediacloud.api.DirectoryApi:
     """
     A central place to get the Media Cloud client
     :return: an media cloud client with the API key from the environment variable
@@ -131,4 +138,12 @@ def get_email_config() -> Dict:
         port=os.environ.get('SMTP_PORT', None),
         from_address=os.environ.get('SMTP_FROM', None),
         notify_emails=os.environ.get('NOTIFY_EMAILS', "").split(",")
+    )
+
+
+def get_slack_config() -> Dict:
+    return dict(
+        app_token=SLACK_APP_TOKEN,
+        bot_token=SLACK_BOT_TOKEN,
+        channel_id=SLACK_CHANNEL_ID,
     )