+++ processor/notifications.py
@@ -2,9 +2,12 @@ import logging
 import smtplib
 import ssl
 from typing import List
-
 from processor import is_email_configured, get_email_config
-
+import time
+from slack_sdk import WebClient
+from slack_sdk.errors import SlackApiError
+import tempfile
+import os
 
 logger = logging.getLogger(__name__)
 
@@ -31,3 +34,43 @@ def send_email(recipients: List[str], subject: str, message: str) -> bool:
             server.sendmail(email_config['from_address'], email_address, msg.encode("utf8"))
     logger.info("  sent")
     return True
+
+
+def upload_to_slack(channel_id: str, bot_key: str, source: str, subject: str, file_path) -> bool:
+    client = WebClient(token=bot_key)
+    try:    
+        filename = source + "-" + time.strftime("%Y%m%d-%H%M%S")
+        response = client.files_upload_v2(
+            channel=channel_id,
+            file=file_path,
+            title=filename.upper(),
+            initial_comment=subject
+        )
+        if response["ok"]:
+            return True
+        else:
+            return False
+    except SlackApiError as e:
+        print(f"Slack API error: {e}")
+        return False
+
+def send_slack_msg(channel_id, bot_key, data_source: str, subject: str, message: str):
+    header = f"{subject.upper()}"
+    formatted_message = f"{header}\n\n{message}"
+    channel = channel_id
+
+    with tempfile.NamedTemporaryFile(mode="w", encoding="utf-8", delete=False) as temp_file:
+        temp_file.write(formatted_message)
+        temp_file.flush()
+        temp_file_name = temp_file.name
+        if upload_to_slack(channel, bot_key, data_source, header, temp_file_name):
+            logger.info("Slack message sent successfully")
+        else:
+            logger.error("Failed to send Slack message")
+
+    if temp_file_name is not None:
+        os.remove(temp_file_name)
+
+
+
+